// types/enums.ts
export enum GlobalTheme {
  // NOTE: DaisyUI has default themes named 'light' and 'dark', but to avoid inheriting those styles, we use 'sun' and 'moon' instead.
  Light = 'sun',
  Dark = 'moon',
  System = 'system',
}

export enum CanvasView {
  TopDown = 'top-down',
  Isometric = 'isometric',
}

export enum CursorStatus {
  Default = 'default',
  Hovered = 'hovered',
  Hidden = 'hidden',
}